+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.15.0] - 2022-03-28
+
+### Added
+
+- Schema for custom headers and custom query options with annotation `JSON.Schema`.
+- Schema for action/function parameters with annotation `JSON.Schema`.
+
 ## [0.14.0] - 2022-03-15
 
 ### Added